requirements.txt
aiogram==2.19
environs==8.0.0
pandas
django
djangorestframework
django-jazzmin